requirements.txt
#Django==4.2.14
#django-bootstrap-v5==1.0.11
#django-environ==0.11.2
#
#mysqlclient==2.1.1
#django-bootstrap-v5
#python-dateutil
#sqlalchemy
#pymysql
wheel
django==5.0.7
#django-bootstrap-v5
django-environ
#
#sqlalchemy
#pymysql
mysqlclient
#mysql-python
#python-dateutil

#mysql-connector
#mysql-connector-python
